extract dictionary to SignatureWrapperStore
