+++ src/TestsExtensions/Extensions/PathExtension.cs
@@ -2,7 +2,7 @@ namespace TestsExtensions.Extensions;
 
 internal static class PathExtension
 {
-    public static string GetJsonFileData(string path)
+    public static string GetJsonFileData(this string path)
     {
         var absolutePath = GetAbsolutePath(path);
         if (!File.Exists(path)) 
@@ -11,7 +11,7 @@ internal static class PathExtension
         return File.ReadAllText(path);
     }
     
-    public static string GetAbsolutePath(string path) => Path.IsPathRooted(path)
+    public static string GetAbsolutePath(this string path) => Path.IsPathRooted(path)
         ? path
         : Path.GetRelativePath(Directory.GetCurrentDirectory(), path);
 }
