+++ src/TestsExtensions/ISignatureWrapper.cs
@@ -1,6 +1,6 @@
 namespace TestsExtensions;
 
-public interface ITestObject
+public interface ISignatureWrapper
 {
     IEnumerable<object[]> Deserialize(string json);
     public string Key { get; }
