--- src/TestsExtensions/ParameterData.cs
@@ -1,3 +1,0 @@
-namespace TestsExtensions;
-
-internal record ParameterData(string? Name, Type Type);
